Fixing dispose code contract to determine success/failure by result vs error
